4/24 이민진 : 기업 평가 토큰작업 완료

+++ src/main/java/devconnect/service/CratingService.java
@@ -26,16 +26,16 @@ public class CratingService {
     private final ProjectRepository projectRepository;
     
     // 기업 평가 등록
-    public boolean cratingWrite(CratingDto cratingDto ){
+    public boolean cratingWrite(CratingDto cratingDto , int loginDno ){
         System.out.println("CratingService.cratingWrite");
         // 입력받은 pno와 dno로 각각 엔티티 조회
-        ProjectEntity pno = projectRepository.findById( cratingDto.getPno() )
+        ProjectEntity projectEntity = projectRepository.findById( cratingDto.getPno() )
                 .orElseThrow( () -> new RuntimeException("프로젝트 번호가 존재하지 않습니다."));
-        DeveloperEntity dno = developerRepository.findById( cratingDto.getDno() )
+        DeveloperEntity developerEntity = developerRepository.findById( cratingDto.getDno() )
                 .orElseThrow( () -> new RuntimeException("기업 번호가 존재하지 않습니다.") );
         // pno , dno 에 값 유무에 따라 true/false 반환
-        if( pno != null && dno != null ){
-            CratingEntity cratingEntity = cratingDto.toEntity(pno,dno);
+        if( projectEntity != null && developerEntity != null && developerEntity.getDno() == loginDno ){
+            CratingEntity cratingEntity = cratingDto.toEntity(projectEntity,developerEntity);
             // ENtity에 저장
             CratingEntity saveEntity = cratingRepository.save(cratingEntity);
             // 결과 반환
@@ -47,43 +47,48 @@ public class CratingService {
     } // f end
     
     // 기업 평가 전체 조회
-    public List<CratingDto> cratingList(){
+    public List<CratingDto> cratingList( int loginDno ){
         System.out.println("CratingService.cratingList");
         // CratingEntity를 Repository를 이용하여 모든 정보 조회
         List<CratingEntity> cratingListAll = cratingRepository.findAll();
         // 필요한 필요한 정보들만 담은 Dto를 반환할 리스트 객체 생성
         List<CratingDto> cratingList = new ArrayList<>();
-        // 조회한 정보의 수 만큼 반복문
-        for( int i = 0; i <= cratingListAll.size() -1; i++ ){
-            // 정보를 담을 Dto객체를 만들고 필요한 정보만 담기
-            CratingDto cratingDto = cratingListAll.get(i).toDto();
-            // 반환할 리스트객체에 하나씩 저장
-            cratingList.add( cratingDto );
-        } // for end
+        // 기업이 로그인 상태일 경우에만
+        if( loginDno >= 1 ) {
+            // 조회한 정보의 수 만큼 반복문
+            for (int i = 0; i <= cratingListAll.size() - 1; i++) {
+                // 정보를 담을 Dto객체를 만들고 필요한 정보만 담기
+                CratingDto cratingDto = cratingListAll.get(i).toDto();
+                // 반환할 리스트객체에 하나씩 저장
+                cratingList.add(cratingDto);
+            } // for end
+        } // if end
         // 리스트 객체 반환
         return cratingList;
     } // f end
 
     // 기업 평가 개별 조회
-    public CratingDto cratingView( int crno ){
+    public CratingDto cratingView( int crno , int loginDno ){
         System.out.println("CratingService.cratingView");
-        // crno 를 기반으로 평가 조회
-        Optional< CratingEntity > optional = cratingRepository.findById( crno );
-        // 값이 있으면 true 없으면 false 로 조건문
-        if( optional.isPresent() ){
-            // 값을 Entity객체에 대입
-            CratingEntity cratingEntity = optional.get();
-            // Dto로 변환
-            CratingDto cratingDto = cratingEntity.toDto();
-            // dto 반환
-            return cratingDto;
+        if ( loginDno >= 1 ) {
+            // crno 를 기반으로 평가 조회
+            Optional<CratingEntity> optional = cratingRepository.findById(crno);
+            // 값이 있으면 true 없으면 false 로 조건문
+            if (optional.isPresent()) {
+                // 값을 Entity객체에 대입
+                CratingEntity cratingEntity = optional.get();
+                // Dto로 변환
+                CratingDto cratingDto = cratingEntity.toDto();
+                // dto 반환
+                return cratingDto;
+            } // if end
         } // if end
         // 값이 없으면 null 반환
         return null;
     } // f end
     
     // 기업 평가 수정
-    public boolean cratingUpdate( CratingDto cratingDto ){
+    public boolean cratingUpdate( CratingDto cratingDto , int loginDno ){
         System.out.println("CratingService.cratingUpdate");
         // dto에 입력한 crno로 해당 엔티티 조회
         Optional<CratingEntity> optional = cratingRepository.findById( cratingDto.getCrno() );
@@ -91,18 +96,21 @@ public class CratingService {
         if( optional.isPresent() ){
             // 값을 Entity 객체에 대입
             CratingEntity cratingEntity = optional.get();
-            // dto에 입력한 값으로 Entity 수정
-            cratingEntity.setCrscore( cratingDto.getCrscore() );
-            cratingEntity.setCrstate( cratingDto.getCrstate() );
-            // 값 수정후 true 반환
-            return true;
+            // 기업을 수정 하는 dno와 기업을 등록했던 dno가 같은지 조건문
+            if( cratingEntity.getDeveloperEntity().getDno() == loginDno ) {
+                // dto에 입력한 값으로 Entity 수정
+                cratingEntity.setCrscore(cratingDto.getCrscore());
+                cratingEntity.setCrstate(cratingDto.getCrstate());
+                // 값 수정후 true 반환
+                return true;
+            } // if end
         } // if end
         // 값이 없으면 false 반환
         return false;
     } // f end
 
     // 기업 평가 삭제
-    public boolean cratingDelete( int crno ){
+    public boolean cratingDelete( int crno , int loginDno ){
         System.out.println("CratingService.cratingDelete");
         // crno에 해당하는 엔티티 조회
         Optional<CratingEntity> optional = cratingRepository.findById( crno );
@@ -110,9 +118,12 @@ public class CratingService {
         if( optional.isPresent() ){
             // 값을 Entity 객체에 대입
             CratingEntity cratingEntity = optional.get();
-            // Entity 객체에서 Drno를 찾아서 그걸 기반으로 데이터삭제
-            cratingRepository.deleteById( cratingEntity.getCrno() );
-            return true;
+            // 평가를 등록했던 개발자와 삭제버튼을 누르는 개발자와 dno가 같은지 확인
+            if( cratingEntity.getDeveloperEntity().getDno() == loginDno ) {
+                // Entity 객체에서 Drno를 찾아서 그걸 기반으로 데이터삭제
+                cratingRepository.deleteById(cratingEntity.getCrno());
+                return true;
+            } // if end
         } // if end
         // 값이 없으면 false 반환
         return false;
